msm: kgsl: Fix the busy counter initialization sequence
The busy counters should be cleared when the GPU finishes starting
up but before it begins processing.  At this point it is not yet
in ACTIVE state, but the counters should be read.  Allow this read
to go through and initialize the counter values.  If this does not
happen then the first returned counter value after SLUMBER will be
zeros.

Change-Id: I9292b536bc9dfd3c905efb152de8834a4029c8bd
Signed-off-by: Lucille Sylvester <lsylvest@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>
Signed-off-by: massstash <massstash@gmail.com>
Signed-off-by: MassStash <MassStash@gmail.com>

Conflicts:
	drivers/gpu/msm/adreno.c